@@ -17,7 +17,7 @@ dependencies:
   cupertino_icons: ^1.0.2
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dev_dependencies:
   test: ^1.16.0
