@@ -12,9 +12,9 @@ class Event {
 class SharedEvents extends InheritedWidget {
   final List<Event> events;
 
-  SharedEvents({this.events, child: Widget}) : super(child: child);
+  SharedEvents({required this.events, child: Widget}) : super(child: child);
 
-  static SharedEvents of(BuildContext context) {
+  static SharedEvents? of(BuildContext context) {
     return context.dependOnInheritedWidgetOfExactType();
   }
 
